requirements.txt
## The following requirements were added by pip freeze:
beautifulsoup4==4.6.0
html5lib==0.999999999
python-telegram-bot==6.0.3
requests==2.11.1